minor, distinguish between redland and non-redland in gandaux(1) upon build
This is purely because the mechanics for renaming the index file(s)
are different.

tcbdb is a one-store/one-file approach and hence renaming the store
is renaming the file and is atomic.

redland is a one-store/multiple-files approach and we use the temporary
file name only as a stamp file, the store is created under the final
name directly and only the stamp file will be renamed.

This is likely to change in the future, however.